Use sourceContents when non-null, even if it's an empty string
When using an optimizing compiler, it may be that all the functions in a
file have been inlined into their call sites, leaving the file empty.
But there are still references to that file in other modules, whose source
maps will contain `sources` and `sourceContents` entries for the empty file.

The `sourceContents` for the empty files is `""`, the empty string.
That's falsy in JavaScript which causes this check to wrongly fail,
leading to an attempt to open the formerly empty but now missing file,
breaking compilation.

This tweaks the check to allow an empty but non-`null` `sourceContents`.